Initialize CVS branch netbsd-2-1
The per-file CVS branch roots do not exactly match a reconstructed source-branch commit.

+++ bin/pax/file_subs.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: file_subs.c,v 1.41 2004/02/13 23:10:14 matt Exp $	*/
+/*	$NetBSD: file_subs.c,v 1.41.2.8.2.1 2005/07/23 17:32:16 snj Exp $	*/
 
 /*-
  * Copyright (c) 1992 Keith Muller.
@@ -42,7 +42,7 @@
 #if 0
 static char sccsid[] = "@(#)file_subs.c	8.1 (Berkeley) 5/31/93";
 #else
-__RCSID("$NetBSD: file_subs.c,v 1.41 2004/02/13 23:10:14 matt Exp $");
+__RCSID("$NetBSD: file_subs.c,v 1.41.2.8.2.1 2005/07/23 17:32:16 snj Exp $");
 #endif
 #endif /* not lint */
 
@@ -67,14 +67,22 @@ char *xtmp_name;
 static int
 mk_link(char *,struct stat *,char *, int);
 
+static int warn_broken;
+
 /*
  * routines that deal with file operations such as: creating, removing;
  * and setting access modes, uid/gid and times of files
  */
+#define SET_BITS		(S_ISUID | S_ISGID)
+#define FILE_BITS		(S_IRWXU | S_IRWXG | S_IRWXO)
+#define A_BITS			(FILE_BITS | SET_BITS | S_ISVTX)
 
-#define FILEBITS		(S_ISVTX | S_IRWXU | S_IRWXG | S_IRWXO)
-#define SETBITS			(S_ISUID | S_ISGID)
-#define ABITS			(FILEBITS | SETBITS)
+/*
+ * The S_ISVTX (sticky bit) can be set by non-superuser on directories
+ * but not other kinds of files.
+ */
+#define FILEBITS(dir)		((dir) ? (FILE_BITS | S_ISVTX) : FILE_BITS)
+#define SETBITS(dir)		((dir) ? SET_BITS : (SET_BITS | S_ISVTX))
 
 /*
  * file_creat()
@@ -84,11 +92,39 @@ mk_link(char *,struct stat *,char *, int);
  */
 
 int
-file_creat(ARCHD *arcn)
+file_creat(ARCHD *arcn, int write_to_hardlink)
 {
 	int fd = -1;
 	int oerrno;
 
+	/*
+	 * Some horribly busted tar implementations, have directory nodes
+	 * that end in a /, but they mark as files. Compensate for that
+	 * by not creating a directory node at this point, but a file node,
+	 * and not creating the temp file.
+	 */
+	if (arcn->nlen != 0 && arcn->name[arcn->nlen - 1] == '/') {
+		if (!warn_broken) {
+			tty_warn(0, "Archive was created with a broken tar;"
+			    " file `%s' is a directory, but marked as plain.",
+			    arcn->name);
+			warn_broken = 1;
+		}
+		return -1;
+	}
+
+	/*
+	 * In "cpio" archives it's usually the last record of a set of
+	 * hardlinks which includes the contents of the file. We cannot
+	 * use a tempory file in that case because we couldn't link it
+	 * with the existing other hardlinks after restoring the contents
+	 * to it. And it's also useless to create the hardlink under a
+	 * temporary name because the other hardlinks would have partial
+	 * contents while restoring.
+	 */
+	if (write_to_hardlink)
+		return (open(arcn->name, O_TRUNC | O_EXCL | O_RDWR, 0));
+	
 	/*
 	 * Create a temporary file name so that the file doesn't have partial
 	 * contents while restoring.
@@ -98,7 +134,7 @@ file_creat(ARCHD *arcn)
 		syswarn(1, errno, "Cannot malloc %d bytes", arcn->nlen + 8);
 		return(-1);
 	}
-	if (xtmp_name)
+	if (xtmp_name != NULL)
 		abort();
 	xtmp_name = arcn->tmp_name;
 
@@ -118,6 +154,7 @@ file_creat(ARCHD *arcn)
 		if (nodirs || chk_path(arcn->name,arcn->sb.st_uid,arcn->sb.st_gid) < 0) {
 			(void)fflush(listf);
 			syswarn(1, oerrno, "Cannot create %s", arcn->tmp_name);
+			xtmp_name = NULL;
 			free(arcn->tmp_name);
 			arcn->tmp_name = NULL;
 			return(-1);
@@ -137,13 +174,17 @@ file_creat(ARCHD *arcn)
 void
 file_close(ARCHD *arcn, int fd)
 {
-	int res = 0;
+	char *tmp_name;
+	int res;
 
 	if (fd < 0)
 		return;
+
+	tmp_name = (arcn->tmp_name != NULL) ? arcn->tmp_name : arcn->name;
+
 	if (close(fd) < 0)
 		syswarn(0, errno, "Cannot close file descriptor on %s",
-		    arcn->tmp_name);
+		    tmp_name);
 
 	/*
 	 * set owner/groups first as this may strip off mode bits we want
@@ -151,7 +192,9 @@ file_close(ARCHD *arcn, int fd)
 	 * modification times.
 	 */
 	if (pids)
-		res = set_ids(arcn->tmp_name, arcn->sb.st_uid, arcn->sb.st_gid);
+		res = set_ids(tmp_name, arcn->sb.st_uid, arcn->sb.st_gid);
+	else
+		res = 0;
 
 	/*
 	 * IMPORTANT SECURITY NOTE:
@@ -159,27 +202,32 @@ file_close(ARCHD *arcn, int fd)
 	 * set uid/gid bits but restore the file modes (since mkstemp doesn't).
 	 */
 	if (!pmode || res)
-		arcn->sb.st_mode &= ~(SETBITS);
+		arcn->sb.st_mode &= ~SETBITS(0);
 	if (pmode)
-		set_pmode(arcn->tmp_name, arcn->sb.st_mode);
+		set_pmode(tmp_name, arcn->sb.st_mode);
 	else
-		set_pmode(arcn->tmp_name, arcn->sb.st_mode & FILEBITS);
+		set_pmode(tmp_name, arcn->sb.st_mode & FILEBITS(0));
 	if (patime || pmtime)
-		set_ftime(arcn->tmp_name, arcn->sb.st_mtime, arcn->sb.st_atime, 0);
-#if HAVE_STRUCT_STAT_ST_FLAGS
-	if (pfflags && arcn->type != PAX_SLK)
-		set_chflags(arcn->tmp_name, arcn->sb.st_flags);
-#endif
+		set_ftime(tmp_name, arcn->sb.st_mtime, arcn->sb.st_atime, 0);
+
+	/* Did we write directly to the target file? */
+	if (arcn->tmp_name == NULL)
+		return;
 
 	/*
 	 * Finally, now the temp file is fully instantiated rename it to
 	 * the desired file name.
 	 */
-	if (rename(arcn->tmp_name, arcn->name) < 0) {
+	if (rename(tmp_name, arcn->name) < 0) {
 		syswarn(0, errno, "Cannot rename %s to %s",
-		    arcn->tmp_name, arcn->name);
-		(void)unlink(arcn->tmp_name);
+		    tmp_name, arcn->name);
+		(void)unlink(tmp_name);
 	}
+
+#if HAVE_STRUCT_STAT_ST_FLAGS
+	if (pfflags && arcn->type != PAX_SLK)
+		set_chflags(arcn->name, arcn->sb.st_flags);
+#endif
 
 	free(arcn->tmp_name);
 	arcn->tmp_name = NULL;
@@ -195,12 +243,21 @@ file_close(ARCHD *arcn, int fd)
  */
 
 int
-lnk_creat(ARCHD *arcn)
+lnk_creat(ARCHD *arcn, int *payload)
 {
 	struct stat sb;
 
 	/*
-	 * we may be running as root, so we have to be sure that link target
+	 * Check if this hardlink carries the "payload". In "cpio" archives
+	 * it's usually the last record of a set of hardlinks which includes
+	 * the contents of the file.
+	 *
+	 */
+	*payload = S_ISREG(arcn->sb.st_mode) &&
+	    (arcn->sb.st_size > 0) && (arcn->sb.st_size <= arcn->skip);
+
+	/*
+	 * We may be running as root, so we have to be sure that link target
 	 * is not a directory, so we lstat and check
 	 */
 	if (lstat(arcn->ln_name, &sb) < 0) {
@@ -314,7 +371,7 @@ mk_link(char *to, struct stat *to_sb, char *from, int ign)
 		/*
 		 * try to get rid of the file, based on the type
 		 */
-		if (S_ISDIR(sb.st_mode)) {
+		if (S_ISDIR(sb.st_mode) && strcmp(from, ".") != 0) {
 			if (rmdir(from) < 0) {
 				syswarn(1, errno, "Cannot remove %s", from);
 				return(-1);
@@ -380,10 +437,10 @@ node_creat(ARCHD *arcn)
 	 * file and link creation routines, this method seems to exhibit the
 	 * best performance in general use workloads.
 	 */
-	file_mode = arcn->sb.st_mode & FILEBITS;
+	file_mode = arcn->sb.st_mode & FILEBITS(arcn->type == PAX_DIR);
 
 	for (;;) {
-		switch(arcn->type) {
+		switch (arcn->type) {
 		case PAX_DIR:
 			/*
 			 * If -h (or -L) was given in tar-mode, follow the
@@ -397,7 +454,8 @@ node_creat(ARCHD *arcn)
 					    sizeof target - 1);
 					if (len == -1) {
 						syswarn(0, errno,
-						   "cannot follow symlink %s in chain for %s",
+						   "cannot follow symlink %s "
+						   "in chain for %s",
 						    nm, arcn->name);
 						res = -1;
 						goto badlink;
@@ -430,7 +488,7 @@ badlink:
 			tty_warn(0,
 			    "%s skipped. Sockets cannot be copied or extracted",
 			    nm);
-			return(-1);
+			return (-1);
 		case PAX_SLK:
 			res = symlink(arcn->ln_name, nm);
 			break;
@@ -444,7 +502,7 @@ badlink:
 			 */
 			tty_warn(0, "%s has an unknown file type, skipping",
 			    nm);
-			return(-1);
+			return (-1);
 		}
 
 		/*
@@ -459,16 +517,27 @@ badlink:
 		 * we failed to make the node
 		 */
 		oerrno = errno;
-		if ((ign = unlnk_exist(nm, arcn->type)) < 0)
-			return(-1);
-
-		if (++pass <= 1)
+		switch (pass++) {
+		case 0:
+			if ((ign = unlnk_exist(nm, arcn->type)) < 0)
+				return (-1);
 			continue;
 
-		if (nodirs || chk_path(nm,arcn->sb.st_uid,arcn->sb.st_gid) < 0) {
-			syswarn(1, oerrno, "Cannot create %s", nm);
-			return(-1);
+		case 1:
+			if (nodirs ||
+			    chk_path(nm, arcn->sb.st_uid,
+			    arcn->sb.st_gid) < 0) {
+				syswarn(1, oerrno, "Cannot create %s", nm);
+				return (-1);
+			}
+			continue;
 		}
+
+		/*
+		 * it must be a file that exists but we can't create or
+		 * remove, but we must avoid the infinite loop.
+		 */
+		break;
 	}
 
 	/*
@@ -485,7 +554,7 @@ badlink:
 	 * set uid/gid bits
 	 */
 	if (!pmode || res)
-		arcn->sb.st_mode &= ~(SETBITS);
+		arcn->sb.st_mode &= ~SETBITS(arcn->type == PAX_DIR);
 	if (pmode)
 		set_pmode(arcn->name, arcn->sb.st_mode);
 
@@ -511,8 +580,9 @@ badlink:
 				 * restored AS CREATED and not as stored if
 				 * pmode is not set.
 				 */
-				set_pmode(nm,
-				    ((sb.st_mode & FILEBITS) | S_IRWXU));
+				set_pmode(nm, ((sb.st_mode &
+				    FILEBITS(arcn->type == PAX_DIR)) |
+				    S_IRWXU));
 				if (!pmode)
 					arcn->sb.st_mode = sb.st_mode;
 			}
@@ -568,8 +638,13 @@ unlnk_exist(char *name, int type)
 	if (S_ISDIR(sb.st_mode)) {
 		/*
 		 * try to remove a directory, if it fails and we were going to
-		 * create a directory anyway, tell the caller (return a 1)
+		 * create a directory anyway, tell the caller (return a 1).
+		 *
+		 * don't try to remove the directory if the name is "."
+		 * otherwise later file/directory creation fails.
 		 */
+		if (strcmp(name, ".") == 0)
+			return(1);
 		if (rmdir(name) < 0) {
 			if (type == PAX_DIR)
 				return(1);
@@ -605,7 +680,7 @@ unlnk_exist(char *name, int type)
  */
 
 int
-chk_path( char *name, uid_t st_uid, gid_t st_gid)
+chk_path(char *name, uid_t st_uid, gid_t st_gid)
 {
 	char *spt = name;
 	struct stat sb;
@@ -668,7 +743,8 @@ chk_path( char *name, uid_t st_uid, gid_t st_gid)
 		 */
 		if ((access(name, R_OK | W_OK | X_OK) < 0) &&
 		    (lstat(name, &sb) == 0)) {
-			set_pmode(name, ((sb.st_mode & FILEBITS) | S_IRWXU));
+			set_pmode(name, ((sb.st_mode & FILEBITS(0)) |
+			    S_IRWXU));
 			add_dir(name, spt - name, &sb, 1);
 		}
 		*(spt++) = '/';
@@ -705,7 +781,7 @@ set_ftime(char *fnm, time_t mtime, time_t atime, int frc)
 		 * set. We get the current values of the times if we need them.
 		 */
 		if (lstat(fnm, &sb) == 0) {
-#ifdef BSD4_4
+#if BSD4_4 && !HAVE_NBTOOL_CONFIG_H
 			if (!patime)
 				TIMESPEC_TO_TIMEVAL(&tv[0], &sb.st_atimespec);
 			if (!pmtime)
@@ -761,7 +837,7 @@ set_ids(char *fnm, uid_t uid, gid_t gid)
 void
 set_pmode(char *fnm, mode_t mode)
 {
-	mode &= ABITS;
+	mode &= A_BITS;
 	if (lchmod(fnm, mode)) {
 		(void)fflush(listf);
 		syswarn(1, errno, "Cannot set permissions on %s", fnm);
@@ -841,6 +917,7 @@ file_write(int fd, char *str, int cnt, int *rem, int *isempt, int sz,
 	int wcnt;
 	char *st = str;
 	char **strp;
+	size_t *lenp;
 
 	/*
 	 * while we have data to process
@@ -900,26 +977,29 @@ file_write(int fd, char *str, int cnt, int *rem, int *isempt, int sz,
 		 * have non-zero data in this file system block, have to write
 		 */
 		switch (fd) {
-		case -1:
+		case -PAX_GLF:
 			strp = &gnu_name_string;
+			lenp = &gnu_name_length;
 			break;
-		case -2:
+		case -PAX_GLL:
 			strp = &gnu_link_string;
+			lenp = &gnu_link_length;
 			break;
 		default:
 			strp = NULL;
+			lenp = NULL;
 			break;
 		}
 		if (strp) {
-			if (*strp)
-				err(1, "WARNING! Major Internal Error! GNU hack Failing!");
-			*strp = malloc(wcnt + 1);
-			if (*strp == NULL) {
+			char *nstr = *strp ? realloc(*strp, *lenp + wcnt + 1) :
+				malloc(wcnt + 1);
+			if (nstr == NULL) {
 				tty_warn(1, "Out of memory");
 				return(-1);
 			}
-			strlcpy(*strp, st, wcnt);
-			break;
+			(void)strlcpy(&nstr[*lenp], st, wcnt + 1);
+			*strp = nstr;
+			*lenp += wcnt;
 		} else if (xwrite(fd, st, wcnt) != wcnt) {
 			syswarn(1, errno, "Failed write to file %s", name);
 			return(-1);
